service/pyproject.toml
[project]
name = "finch"
version = "0.1.0"
description = ""
authors = [{name = "Ben Weems", email = "bencweems@gmail.com"}]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.4",
    "djangorestframework>=3.15",
    "openai>=1.0",
    "python-dotenv>=1.0",
]
